config: Initial implementation
A lot of the fields added are not implemented, but it can configure
the address and port it will bind to, and what endpoints will be
enabled. There's also a struct for serving some of the data provided,
but for now it's not that different from what there was before, other
than the new base_url field.