front: Add Content-Type header to requests
Previously Shoelace would return requests on the frontend as plaintext,
leaving the user's browser or the reverse proxy to interpret the request
as its respective content type. This caused issues with other reverse
proxies such as Caddy. This fixes it by explicitly returning the
Content-Type header. Thanks to the fine folks at projectsegfau.lt for
noting this.