templates/error: Reset line height in description
Having this property caused certain error descriptions to overlap on
each line break. The property was added under the impression it would
reduce margins and make the page more aesthetically pleasing. This was
solved with the `margins` property, so remove the property, and align
the description to the center of the container using `text-align`.