meta: Bump version to v0.2.0
Major release. This release focuses on expanding quite a bit from the
base established in v0.2.x. Changes include:

- RSS feeds for profiles: Users can now fetch the latest posts from
  their favorite creators using an RSS reader.
- Open Graph previews: Posts and profiles can now have embedded previews
  when being shared on other social media platforms.
- Image thumbnails: Pages now load in less time, by fetching lighter
  versions of embedded media, which is now an option in spools v0.6.
- Frontend changes: aoife added the quick jump bar found in the homepage
  to the error pages, reducing the amount of clicks needed to find
  another profile.
- API revamp: Gone are the unnecessary URL formatted headers. The API
  now parses arguments through the path specified in the URL.

It must be noted that RocksDB support has been dropped in this version.
Existing instance administrators are strongly advised to switch to
either a Redis/Redict install, or to use the internal keystore.