license: Relicense under AGPL-v3.0
This might be a bit of a controversial commit, but I'll be explaining
why I took this decision. GPL-v3.0 is not really designed to work on
networking services, like web applications. In my ignorance I used it as
the license for this software, but that wasn't the best decision in
hindsight. To be clear, anything before this commit (so, v0.2.1 and the
Askama rewrite) will keep their GPL-v3.0 license.