main: Implement custom logger
This was probably one of the hardest parts of this port. The tower-http
crate does include the TraceLayer middleware, but it's a little limited
in terms of customization, and splits everything into multiline outputs
instead of using a oneliner approach. Therefore, I had to implement my
custom logger as a function. I might split it off into another function,
since it seems to be cluttering the current main function quite a bit.