src: Revamp crate redundancies
A lot of components in Shoelace were set to be on "crate" visibility.
This is not necessary, since Shoelace is not declared as a library, but
as an application, therefore it's redundant to say its components should
be limited to the crate, since they already are.

Additionally, this commit changes the base_url field in Base to just
"url", for similar reasons as to why the previous change was made, and
removes a pointless response clone in the logger function.